+++ esphome/core/defines.h
@@ -278,6 +278,12 @@
 #define USE_ETHERNET_JL1101
 #define USE_ETHERNET_KSZ8081
 #define USE_ETHERNET_LAN8670
+#define USE_ETHERNET_SPI
+#define USE_ETHERNET_SPI_POLLING_SUPPORT
+#define USE_ETHERNET_OPENETH
+#define CONFIG_ETH_SPI_ETHERNET_W5500 1
+#define CONFIG_ETH_SPI_ETHERNET_DM9051 1
+#define CONFIG_ETH_USE_ESP32_EMAC 1
 #define USE_ETHERNET_MANUAL_IP
 #define USE_ETHERNET_IP_STATE_LISTENERS
 #define USE_ETHERNET_CONNECT_TRIGGER
